// server/src/modules/progress/courseProgress.enums.ts
export enum LectureStatus {
  NOT_STARTED = "not_started",
  IN_PROGRESS = "in_progress",
  COMPLETED = "completed",
}

export enum ProgressMilestone {
  STARTED = "started", // 0%
  QUARTER = "quarter", // 25%
  HALF = "half", // 50%
  THREE_QUARTER = "three_quarter", // 75%
  COMPLETED = "completed", // 100%
}